backend/requirements.txt
fastapi
uvicorn
pydantic
firebase-admin
openai
python-multipart
python-dotenv
ollama

